Add validation messages for ProyectController.

+++ ABM/Controllers/ProyectController.cs
@@ -25,14 +25,22 @@ namespace ABM.Controllers
         [AuthorizeUser(new int[] { administrador })]
         public ActionResult Index()
         {
-            var projects = from p in _projectRepository.GetActiveProjects()
+            
+                var projects = from p in _projectRepository.GetActiveProjects()
                            select new ProyectViewModel()
                            {
                                Id = p.id,
                                ProjectName = p.proyectName,
                                ProjectDetail = p.proyectDetail
                            };
-
+            if (TempData["Error"] != null)
+            {
+                ViewBag.Error = TempData["Error"];
+            }
+            else if(TempData["SucessMessage"] != null)
+            {
+                ViewBag.Message = TempData["SucessMessage"];
+            }
             return View(projects);
         }
 
@@ -40,10 +48,23 @@ namespace ABM.Controllers
         [AuthorizeUser(new int[] { administrador })]
         public ActionResult Details(int id)
         {
-            var model = _projectRepository.GetById(id);
-            ProyectViewModel project = new ProyectViewModel();
-                project.ToProyectViewModel(model);
-            return View(project);
+            try
+            {
+                var model = _projectRepository.GetById(id);
+                if (model != null)
+                {
+                    ProyectViewModel project = new ProyectViewModel();
+                    project.ToProyectViewModel(model);
+                    return View(project);
+                }
+                throw new Exception();
+            }
+            catch (Exception)
+            {
+                TempData["Error"] = "El proyecto es invalido, se redirigio a la pantalla principal.";
+                return RedirectToAction("Index");
+            }
+
         }
 
         // GET: Proyect/Create
@@ -65,13 +86,16 @@ namespace ABM.Controllers
                     var user = (Models.User)Session["User"];
                     model.UserId = user.id;
                     _projectRepository.InsertProject(model.ToEntity());
+                    TempData["SucessMessage"] = "El proyecto fue creado con exito.";
+                    return RedirectToAction("Index");
                 }
+                throw new Exception();
 
-                return RedirectToAction("Index");
             }
             catch
             {
-                return View();
+                TempData["Error"] = "Operacion invalida, se redirigio a la pantalla principal.";
+                return RedirectToAction("Index");
             }
         }
 
@@ -79,15 +103,29 @@ namespace ABM.Controllers
         [AuthorizeUser(new int[] { administrador })]
         public ActionResult Edit(int id)
         {
-            var p = _projectRepository.GetById(id);
-            ProyectViewModel viewModel = new ProyectViewModel()
+            try
             {
-                Id = p.id,
-                ProjectName = p.proyectName,
-                ProjectDetail = p.proyectDetail
-            };
+                var p = _projectRepository.GetById(id);
+                if (p != null)
+                {
+                    ProyectViewModel viewModel = new ProyectViewModel()
+                    {
+                        Id = p.id,
+                        ProjectName = p.proyectName,
+                        ProjectDetail = p.proyectDetail
+                    };
+
+                    return View(viewModel);
+                }
+                throw new Exception();
+            }
+            catch (Exception)
+            {
+                TempData["Error"] = "Proyecto invalido, se redirigio a la pantalla principal.";
+                return RedirectToAction("Index");
+            }
             
-            return View(viewModel);
+            
         }
 
         // POST: Proyect/Edit/5
@@ -103,11 +141,15 @@ namespace ABM.Controllers
                     var user = (Models.User)Session["User"];
                     model.UserId = user.id;
                     _projectRepository.UpdateProject(model.ToEntity());
+                    TempData["SucessMessage"] = "El proyecto fue editado con exito.";
+                    return RedirectToAction("Index");
+
                 }
-                return RedirectToAction("Index");
+                throw new Exception();
             }
             catch
             {
+                TempData["Error"] = "Hubo un error al editar el proyecto.";
                 return View();
             }
         }
@@ -116,9 +158,24 @@ namespace ABM.Controllers
         [AuthorizeUser(new int[] { administrador })]
         public ActionResult Delete(int id)
         {
-            _projectRepository.DeleteProject(id);
-            ModelState.Clear();
-            return RedirectToAction("Index", "Proyect");
+            try
+            {
+                //VERIFICA QUE LA ID SEA VALIDA Y EXISTA UN USARIO CON LA MISMA. QUIZAS NO SEA NECESARIO?
+                //REVISAR DESPUES CON PABLO
+                if (_projectRepository.GetById(id) != null)
+                {
+                    _projectRepository.DeleteProject(id);
+                    ModelState.Clear();
+                    TempData["SucessMessage"] = "El proyecto fue eliminado con exito.";
+                    return RedirectToAction("Index", "Proyect");
+                }
+                throw new Exception();
+            }
+            catch (Exception)
+            {
+                TempData["Error"] = "Hubo un error al eliminar el proyecto.";
+                return View();
+            }
         }
     }
 }
